Extract reeusable parts from actions step
