+++ src/modules/governance/components/createExecuteActionsForm/createExecuteActionsFormActions.tsx
@@ -1,23 +1,13 @@
-import {
-    type IProposalActionsArrayControls,
-    type ProposalActionComponent,
-    ProposalActions,
-} from '@aragon/gov-ui-kit';
-import { useCallback, useEffect } from 'react';
-import { useFieldArray, useFormContext, useWatch } from 'react-hook-form';
-import { ProposalActionType } from '@/modules/governance/api/governanceService';
+import type { ProposalActionComponent } from '@aragon/gov-ui-kit';
+import { useEffect } from 'react';
 import { useDao, useDaoPermissions } from '@/shared/api/daoService';
 import { useTranslations } from '@/shared/components/translationsProvider';
 import { useDaoChain } from '@/shared/hooks/useDaoChain';
-import { proposalActionUtils } from '../../utils/proposalActionUtils';
+import { useProposalActionsField } from '../../hooks/useProposalActionsField';
 import { ActionComposer, actionComposerUtils } from '../actionComposer';
-import type {
-    ICreateProposalFormData,
-    IProposalActionData,
-} from '../createProposalForm';
-import { TransferAssetAction } from '../createProposalForm/createProposalFormActions/proposalActions/transferAssetAction';
-import { UpdateDaoMetadataAction } from '../createProposalForm/createProposalFormActions/proposalActions/updateDaoMetadataAction';
-import { UpdatePluginMetadataAction } from '../createProposalForm/createProposalFormActions/proposalActions/updatePluginMetadataAction';
+import type { IProposalActionData } from '../createProposalForm';
+import { coreCustomActionComponents } from '../createProposalForm';
+import { ProposalActionsEditList } from '../proposalActionsEditList';
 
 export interface ICreateExecuteActionsFormActionsProps {
     /**
@@ -25,16 +15,6 @@ export interface ICreateExecuteActionsFormActionsProps {
      */
     daoId: string;
 }
-
-const coreCustomActionComponents: Record<
-    string,
-    ProposalActionComponent<IProposalActionData>
-> = {
-    [ProposalActionType.TRANSFER]: TransferAssetAction,
-    [actionComposerUtils.transferActionLocked]: TransferAssetAction,
-    [ProposalActionType.METADATA_UPDATE]: UpdateDaoMetadataAction,
-    [ProposalActionType.METADATA_PLUGIN_UPDATE]: UpdatePluginMetadataAction,
-};
 
 export const CreateExecuteActionsFormActions: React.FC<
     ICreateExecuteActionsFormActionsProps
@@ -45,29 +25,15 @@ export const CreateExecuteActionsFormActions: React.FC<
     const { data: dao } = useDao({ urlParams: { id: daoId } });
     const { chainId } = useDaoChain({ daoId });
 
-    const { control, getValues, setValue } =
-        useFormContext<ICreateProposalFormData>();
-
     const {
-        fields: actions,
-        append,
-        remove,
-    } = useFieldArray({ control, name: 'actions' });
+        actions,
+        actionsMerged,
+        handleAddAction,
+        handleRemoveAllActions,
+        getArrayControls,
+    } = useProposalActionsField();
 
-    // We need to watch because action views can update data, and it's not reflected otherwise!
-    // We merge it with `actions` because of `id` and other internal props which are missing in watched action.
-    const watchActions = useWatch<
-        Record<string, ICreateProposalFormData['actions']>
-    >({ name: 'actions' });
-    // Skip stale watch data when lengths diverge after remove() to avoid index corruption.
-    const stableWatchActions =
-        watchActions?.length === actions.length ? watchActions : undefined;
-    const actionsMerged = actions.map((field, index) => ({
-        ...field,
-        ...stableWatchActions?.[index],
-        id: field.id,
-    }));
-
+    // TODO: implement useAllDaoPermissions
     const {
         data: daoPermissionsData,
         hasNextPage,
@@ -88,70 +54,6 @@ export const CreateExecuteActionsFormActions: React.FC<
         (page) => page.data,
     );
 
-    /**
-     * Note: We don't use useFieldArray.swap() or .move() because they create empty slots
-     * when dealing with complex nested objects, causing data loss and crashes. Instead,
-     * we use structuredClone to create a deep copy, manually swap elements, and update
-     * the entire array at once.
-     */
-    const handleMoveAction = useCallback(
-        (index: number, newIndex: number) => {
-            if (newIndex < 0 || newIndex >= actions.length) {
-                return;
-            }
-
-            const currentActions = getValues('actions');
-            const actionsCopy = structuredClone(currentActions);
-
-            const temp = actionsCopy[index];
-            actionsCopy[index] = actionsCopy[newIndex];
-            actionsCopy[newIndex] = temp;
-
-            setValue('actions', actionsCopy, {
-                shouldValidate: false,
-                shouldDirty: true,
-                shouldTouch: false,
-            });
-        },
-        [actions, getValues, setValue],
-    );
-
-    const handleRemoveAction = (index: number) => {
-        remove(index);
-    };
-
-    const handleAddAction = (newActions: IProposalActionData[]) => {
-        append(newActions);
-    };
-
-    const handleRemoveAllActions = useCallback(() => {
-        remove();
-    }, [remove]);
-
-    const getArrayControls = (
-        index: number,
-    ): IProposalActionsArrayControls<IProposalActionData> => ({
-        moveUp: {
-            label: t('app.governance.createProposalForm.actions.editAction.up'),
-            onClick: (index) => handleMoveAction(index, index - 1),
-            disabled: actions.length < 2 || index === 0,
-        },
-        moveDown: {
-            label: t(
-                'app.governance.createProposalForm.actions.editAction.down',
-            ),
-            onClick: (index) => handleMoveAction(index, index + 1),
-            disabled: actions.length < 2 || index === actions.length - 1,
-        },
-        remove: {
-            label: t(
-                'app.governance.createProposalForm.actions.editAction.remove',
-            ),
-            onClick: handleRemoveAction,
-            disabled: false,
-        },
-    });
-
     const { pluginComponents } = actionComposerUtils.getDaoPluginActions(dao);
     const { components: permissionActionComponents } =
         actionComposerUtils.getDaoPermissionActions({
@@ -169,39 +71,17 @@ export const CreateExecuteActionsFormActions: React.FC<
     };
 
     const showActionComposer = dao != null;
+    // TODO: is actions relaly needed? can we use mergedActions?
     const hasActions = actions.length > 0;
-
-    const expandedActions = actions.map((action) => action.id);
-    const noOpActionsChange = useCallback(() => undefined, []);
 
     return (
         <div className="flex flex-col gap-y-10">
-            <ProposalActions.Root
-                editMode={true}
-                expandedActions={expandedActions}
-                onExpandedActionsChange={noOpActionsChange}
-            >
-                <ProposalActions.Container emptyStateDescription="">
-                    {actionsMerged.map((action, index) => (
-                        <ProposalActions.Item<IProposalActionData>
-                            action={action as IProposalActionData}
-                            actionCount={actionsMerged.length}
-                            actionFunctionSelector={proposalActionUtils.actionToFunctionSelector(
-                                action as IProposalActionData,
-                            )}
-                            arrayControls={getArrayControls(index)}
-                            CustomComponent={
-                                customActionComponents[action.type]
-                            }
-                            chainId={chainId}
-                            editMode={true}
-                            formPrefix={`actions.${index.toString()}`}
-                            key={action.id}
-                            value={action.id}
-                        />
-                    ))}
-                </ProposalActions.Container>
-            </ProposalActions.Root>
+            <ProposalActionsEditList
+                actionsMerged={actionsMerged}
+                chainId={chainId}
+                customActionComponents={customActionComponents}
+                getArrayControls={getArrayControls}
+            />
             {showActionComposer ? (
                 <ActionComposer
                     daoId={daoId}
