+++ src/modules/governance/components/createProposalForm/createProposalFormActions/createProposalFormActions.tsx
@@ -1,34 +1,25 @@
-import {
-    addressUtils,
-    type IProposalActionsArrayControls,
-    type ProposalActionComponent,
-    ProposalActions,
-} from '@aragon/gov-ui-kit';
+import { addressUtils, type ProposalActionComponent } from '@aragon/gov-ui-kit';
 import { useCallback, useEffect, useState } from 'react';
-import { useFieldArray, useFormContext, useWatch } from 'react-hook-form';
+import { useFormContext } from 'react-hook-form';
 import { useAllowedActions } from '@/modules/governance/api/executeSelectorsService';
-import {
-    type IProposalAction,
-    ProposalActionType,
-} from '@/modules/governance/api/governanceService';
+import type { IProposalAction } from '@/modules/governance/api/governanceService';
 import { useDao, useDaoPermissions } from '@/shared/api/daoService';
 import { useTranslations } from '@/shared/components/translationsProvider';
 import { useDaoChain } from '@/shared/hooks/useDaoChain';
 import { daoUtils } from '@/shared/utils/daoUtils';
 import { monitoringUtils } from '@/shared/utils/monitoringUtils';
 import type { IProposalCreateAction } from '../../../dialogs/publishProposalDialog';
+import { useProposalActionsField } from '../../../hooks/useProposalActionsField';
 import { proposalActionPreparationUtils } from '../../../utils/proposalActionPreparationUtils';
 import { proposalActionsImportExportUtils } from '../../../utils/proposalActionsImportExportUtils';
-import { proposalActionUtils } from '../../../utils/proposalActionUtils';
 import { ActionComposer, actionComposerUtils } from '../../actionComposer';
+import { ProposalActionsEditList } from '../../proposalActionsEditList';
 import type {
     ICreateProposalFormData,
     IProposalActionData,
 } from '../createProposalFormDefinitions';
 import { useCreateProposalFormContext } from '../createProposalFormProvider';
-import { TransferAssetAction } from './proposalActions/transferAssetAction';
-import { UpdateDaoMetadataAction } from './proposalActions/updateDaoMetadataAction';
-import { UpdatePluginMetadataAction } from './proposalActions/updatePluginMetadataAction';
+import { coreCustomActionComponents } from './coreCustomActionComponents';
 
 export interface ICreateProposalFormActionsProps {
     /**
@@ -40,16 +31,6 @@ export interface ICreateProposalFormActionsProps {
      */
     pluginAddress: string;
 }
-
-const coreCustomActionComponents: Record<
-    string,
-    ProposalActionComponent<IProposalActionData>
-> = {
-    [ProposalActionType.TRANSFER]: TransferAssetAction,
-    [actionComposerUtils.transferActionLocked]: TransferAssetAction,
-    [ProposalActionType.METADATA_UPDATE]: UpdateDaoMetadataAction,
-    [ProposalActionType.METADATA_PLUGIN_UPDATE]: UpdatePluginMetadataAction,
-};
 
 export const CreateProposalFormActions: React.FC<
     ICreateProposalFormActionsProps
@@ -83,33 +64,17 @@ export const CreateProposalFormActions: React.FC<
     const { t } = useTranslations();
     const { chainId } = useDaoChain({ daoId });
 
-    const { control, getValues, setValue } =
-        useFormContext<ICreateProposalFormData>();
+    const { getValues } = useFormContext<ICreateProposalFormData>();
 
     const { prepareActions } = useCreateProposalFormContext();
 
     const {
-        fields: actions,
-        append,
-        remove,
-    } = useFieldArray({
-        control,
-        name: 'actions',
-    });
-
-    // We need to watch because action views can update data, and it's not reflected otherwise!
-    // We merge it with `actions` because of `id` and other internal props which are missing in watched action.
-    const watchActions = useWatch<
-        Record<string, ICreateProposalFormData['actions']>
-    >({ name: 'actions' });
-    // Skip stale watch data when lengths diverge after remove() to avoid index corruption.
-    const stableWatchActions =
-        watchActions?.length === actions.length ? watchActions : undefined;
-    const actionsMerged = actions.map((field, index) => ({
-        ...field,
-        ...stableWatchActions?.[index],
-        id: field.id,
-    }));
+        actions,
+        actionsMerged,
+        handleAddAction,
+        handleRemoveAllActions,
+        getArrayControls,
+    } = useProposalActionsField();
 
     const { data: allowedActionsData } = useAllowedActions(
         {
@@ -144,46 +109,6 @@ export const CreateProposalFormActions: React.FC<
     const [isDownloadPinning, setIsDownloadPinning] = useState(false);
     const [hasDownloadPinErrors, setHasDownloadPinErrors] = useState(false);
 
-    /**
-     * Note: We don't use useFieldArray.swap() or .move() because they create empty slots
-     * when dealing with complex nested objects, causing data loss and crashes. Instead,
-     * we use structuredClone to create a deep copy, manually swap elements, and update
-     * the entire array at once.
-     */
-    const handleMoveAction = useCallback(
-        (index: number, newIndex: number) => {
-            if (newIndex < 0 || newIndex >= actions.length) {
-                return;
-            }
-
-            const currentActions = getValues('actions');
-            const actionsCopy = structuredClone(currentActions);
-
-            const temp = actionsCopy[index];
-            actionsCopy[index] = actionsCopy[newIndex];
-            actionsCopy[newIndex] = temp;
-
-            setValue('actions', actionsCopy, {
-                shouldValidate: false,
-                shouldDirty: true,
-                shouldTouch: false,
-            });
-        },
-        [actions, getValues, setValue],
-    );
-
-    const handleRemoveAction = (index: number) => {
-        remove(index);
-    };
-
-    const handleAddAction = (newActions: IProposalActionData[]) => {
-        append(newActions);
-    };
-
-    const handleRemoveAllActions = useCallback(() => {
-        remove();
-    }, [remove]);
-
     const handleDownloadActions = useCallback(async () => {
         setIsDownloadPinning(true);
         setHasDownloadPinErrors(false);
@@ -216,30 +141,6 @@ export const CreateProposalFormActions: React.FC<
         }
     }, [daoId, getValues, prepareActions]);
 
-    const getArrayControls = (
-        index: number,
-    ): IProposalActionsArrayControls<IProposalActionData> => ({
-        moveUp: {
-            label: t('app.governance.createProposalForm.actions.editAction.up'),
-            onClick: (index) => handleMoveAction(index, index - 1),
-            disabled: actions.length < 2 || index === 0,
-        },
-        moveDown: {
-            label: t(
-                'app.governance.createProposalForm.actions.editAction.down',
-            ),
-            onClick: (index) => handleMoveAction(index, index + 1),
-            disabled: actions.length < 2 || index === actions.length - 1,
-        },
-        remove: {
-            label: t(
-                'app.governance.createProposalForm.actions.editAction.remove',
-            ),
-            onClick: handleRemoveAction,
-            disabled: false,
-        },
-    });
-
     const { pluginComponents } = actionComposerUtils.getDaoPluginActions(dao);
     const { components: permissionActionComponents } =
         actionComposerUtils.getDaoPermissionActions({
@@ -261,37 +162,14 @@ export const CreateProposalFormActions: React.FC<
         targetDao != null;
     const hasActions = actions.length > 0;
 
-    const expandedActions = actions.map((action) => action.id);
-    const noOpActionsChange = useCallback(() => undefined, []);
-
     return (
         <div className="flex flex-col gap-y-10">
-            <ProposalActions.Root
-                editMode={true}
-                expandedActions={expandedActions}
-                onExpandedActionsChange={noOpActionsChange}
-            >
-                <ProposalActions.Container emptyStateDescription="">
-                    {actionsMerged.map((action, index) => (
-                        <ProposalActions.Item<IProposalActionData>
-                            action={action as IProposalActionData}
-                            actionCount={actionsMerged.length}
-                            actionFunctionSelector={proposalActionUtils.actionToFunctionSelector(
-                                action as IProposalActionData,
-                            )}
-                            arrayControls={getArrayControls(index)}
-                            CustomComponent={
-                                customActionComponents[action.type]
-                            }
-                            chainId={chainId}
-                            editMode={true}
-                            formPrefix={`actions.${index.toString()}`}
-                            key={action.id}
-                            value={action.id}
-                        />
-                    ))}
-                </ProposalActions.Container>
-            </ProposalActions.Root>
+            <ProposalActionsEditList
+                actionsMerged={actionsMerged}
+                chainId={chainId}
+                customActionComponents={customActionComponents}
+                getArrayControls={getArrayControls}
+            />
             {showActionComposer ? (
                 <ActionComposer
                     allowedActions={allowedActions}
