+++ src/modules/governance/components/createProposalForm/createProposalFormActions/index.ts
@@ -1,3 +1,4 @@
+export { coreCustomActionComponents } from './coreCustomActionComponents';
 export {
     CreateProposalFormActions,
     type ICreateProposalFormActionsProps,
