+++ src/modules/governance/components/proposalActionsEditList/index.ts
@@ -1,0 +1,4 @@
+export {
+    type IProposalActionsEditListProps,
+    ProposalActionsEditList,
+} from './proposalActionsEditList';
