+++ src/modules/governance/hooks/useProposalActionsField/index.ts
@@ -1,0 +1,1 @@
+export { useProposalActionsField } from './useProposalActionsField';
